backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.4
pydantic-settings==2.3.0
bcrypt==4.1.3
PyJWT==2.8.0
supabase==2.5.0
google-generativeai==0.8.3
openai==1.35.0
slowapi==0.1.9
python-multipart==0.0.9
